requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
appdirs==1.4.4
attrs==21.4.0
bidict==0.21.4
branca==0.4.2
caer==2.0.8
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
folium==0.12.1.post1
fonttools==4.29.1
google-trans-new==1.1.9
googletrans==4.0.0rc1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.16.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
langdetect==1.0.9
MarkupSafe==2.0.1
matplotlib==3.5.1
mediapipe==0.8.9.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.22.0
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.3.5
Pillow==9.0.1
protobuf==3.19.4
pyparsing==3.0.7
pyshine==0.0.9
pytesseract==0.3.8
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
python-utils==3.1.0
pytz==2021.3
requests==2.27.1
rfc3986==1.5.0
simple-websocket==0.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
tesseract==0.1.3
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
utils==1.0.1
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.1
